Raspberry_Pi4/modbus_bridge/requirements.txt
pymodbus==2.5.3
pyserial